fix: Небольшие исправления
